Google Docs and a static Job Manager
