+++ gpslogger/src/main/java/com/mendhak/gpslogger/common/events/GDocsEvent.java
@@ -1,0 +1,9 @@
+package com.mendhak.gpslogger.common.events;
+
+
+public class GDocsEvent {
+    public boolean success;
+    public GDocsEvent(boolean success){
+        this.success = success;
+    }
+}
